+++ src/main/java/com/fein91/dao/InvoiceRepository.java
@@ -4,30 +4,28 @@ import com.fein91.model.Counterparty;
 import com.fein91.model.Invoice;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
-import org.springframework.data.repository.Repository;
 import org.springframework.data.repository.query.Param;
 
-import java.math.BigInteger;
 import java.util.List;
 
 /**
- * Created by olta1014 on 23.05.2016.
+ * Invoice repository
  */
-public interface InvoiceRepository extends CrudRepository<Invoice, BigInteger> {
+public interface InvoiceRepository extends CrudRepository<Invoice, Long> {
 
     List<Invoice> findBySource(Counterparty counterparty);
 
     List<Invoice> findByTarget(Counterparty counterparty);
 
     @Query("SELECT i FROM Invoice i where i.source.id = :counterPartyId")
-    List<Invoice> findInvoicesBySourceId(@Param("counterPartyId") BigInteger counterPartyId);
+    List<Invoice> findInvoicesBySourceId(@Param("counterPartyId") Long counterPartyId);
 
     @Query("SELECT i FROM Invoice i where i.target.id = :counterPartyId")
-    List<Invoice> findInvoicesByTargetId(@Param("counterPartyId") BigInteger counterPartyId);
+    List<Invoice> findInvoicesByTargetId(@Param("counterPartyId") Long counterPartyId);
 
     List<Invoice> findBySourceAndTarget(Counterparty source, Counterparty target);
 
     @Query("SELECT i FROM Invoice i where i.target.id = :targetId and i.source.id = :sourceId")
-    List<Invoice> findBySourceAndTarget(@Param("sourceId") BigInteger sourceId, @Param("targetId") BigInteger targetId);
+    List<Invoice> findBySourceAndTarget(@Param("sourceId") Long sourceId, @Param("targetId") Long targetId);
 
 }
